requirements.txt
anyio==3.6.2
click==8.1.3
fastapi==0.95.2
h11==0.14.0
httptools==0.5.0
idna==3.4
pydantic==1.10.7
python-dotenv==1.0.0
PyYAML==6.0
sniffio==1.3.0
starlette==0.27.0
typing-extensions==4.5.0
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
Pillow==9.5.0
pymongo==4.3.3
python-multipart==0.0.6
DeepImageSearch
